backend/requirements.txt
django
djangorestframework
psycopg2-binary
django-environ
djangorestframework-simplejwt
django-cors-headers


